CORDA-4021: Remove last reference to ASM from code package.
