+++ djvm/src/main/kotlin/net/corda/djvm/analysis/impl/JavaAnnotationConfiguration.kt
@@ -2,7 +2,7 @@
 package net.corda.djvm.analysis.impl
 
 import net.corda.djvm.analysis.AnalysisConfiguration.Companion.sandboxed
-import net.corda.djvm.code.EmitterModule
+import net.corda.djvm.code.impl.EmitterModuleImpl
 import net.corda.djvm.references.Member
 import org.objectweb.asm.Opcodes.*
 
@@ -23,7 +23,7 @@ fun generateJavaAnnotationMethods(): List<Member> = listOf(
         memberName = "jvmAnnotation",
         descriptor = "()Ljava/lang/annotation/Annotation;"
     ) {
-        override fun writeBody(emitter: EmitterModule) = with(emitter) {
+        override fun writeBody(emitter: EmitterModuleImpl) = with(emitter) {
             pushNull()
             returnObject()
         }
