+++ djvm/src/main/kotlin/net/corda/djvm/analysis/impl/JavaNioBitsConfiguration.kt
@@ -1,9 +1,9 @@
 @file:JvmName("JavaNioBitsConfiguration")
 package net.corda.djvm.analysis.impl
 
-import net.corda.djvm.code.EmitterModule
 import net.corda.djvm.code.impl.CLASS_CONSTRUCTOR_NAME
 import net.corda.djvm.code.impl.DJVM_NAME
+import net.corda.djvm.code.impl.EmitterModuleImpl
 import net.corda.djvm.references.Member
 import org.objectweb.asm.Opcodes.*
 
@@ -21,7 +21,7 @@ fun generateJavaBitsMethods(): List<Member> = listOf(
          * Replace <clinit>():
          *     byteOrder = DJVM.getNativeOrder()
          */
-        override fun writeBody(emitter: EmitterModule) = with(emitter) {
+        override fun writeBody(emitter: EmitterModuleImpl) = with(emitter) {
             invokeStatic(DJVM_NAME, "getNativeOrder", "()Lsandbox/java/nio/ByteOrder;")
             putStatic(className, "byteOrder", "Lsandbox/java/nio/ByteOrder;")
             returnVoid()
