+++ djvm/src/main/kotlin/net/corda/djvm/analysis/impl/JavaResourceBundleConfiguration.kt
@@ -2,8 +2,8 @@
 package net.corda.djvm.analysis.impl
 
 import net.corda.djvm.analysis.AnalysisConfiguration.Companion.sandboxed
-import net.corda.djvm.code.EmitterModule
 import net.corda.djvm.code.impl.DJVM_NAME
+import net.corda.djvm.code.impl.EmitterModuleImpl
 import net.corda.djvm.references.Member
 import org.objectweb.asm.Label
 import org.objectweb.asm.Opcodes.*
@@ -25,7 +25,7 @@ fun generateJavaResourceBundleMethods(): List<Member> = listOf(
         memberName = GET_BUNDLE,
         descriptor = "(Lsandbox/java/lang/String;)Lsandbox/java/util/ResourceBundle;"
     ) {
-        override fun writeBody(emitter: EmitterModule) = with(emitter) {
+        override fun writeBody(emitter: EmitterModuleImpl) = with(emitter) {
             pushObject(0)
             pushDefaultLocale()
             pushDefaultControl()
@@ -40,7 +40,7 @@ fun generateJavaResourceBundleMethods(): List<Member> = listOf(
         memberName = GET_BUNDLE,
         descriptor = "(Lsandbox/java/lang/String;Lsandbox/java/util/ResourceBundle\$Control;)Lsandbox/java/util/ResourceBundle;"
     ) {
-        override fun writeBody(emitter: EmitterModule) = with(emitter) {
+        override fun writeBody(emitter: EmitterModuleImpl) = with(emitter) {
             pushObject(0)
             pushDefaultLocale()
             pushObject(1)
@@ -55,7 +55,7 @@ fun generateJavaResourceBundleMethods(): List<Member> = listOf(
         memberName = GET_BUNDLE,
         descriptor = "(Lsandbox/java/lang/String;Lsandbox/java/util/Locale;)Lsandbox/java/util/ResourceBundle;"
     ) {
-        override fun writeBody(emitter: EmitterModule) = with(emitter) {
+        override fun writeBody(emitter: EmitterModuleImpl) = with(emitter) {
             pushObject(0)
             pushObject(1)
             pushDefaultControl()
@@ -70,7 +70,7 @@ fun generateJavaResourceBundleMethods(): List<Member> = listOf(
         memberName = GET_BUNDLE,
         descriptor = "(Lsandbox/java/lang/String;Lsandbox/java/util/Locale;Lsandbox/java/util/ResourceBundle\$Control;)Lsandbox/java/util/ResourceBundle;"
     ) {
-        override fun writeBody(emitter: EmitterModule) = with(emitter) {
+        override fun writeBody(emitter: EmitterModuleImpl) = with(emitter) {
             pushObject(0)
             pushObject(1)
             pushObject(2)
@@ -85,7 +85,7 @@ fun generateJavaResourceBundleMethods(): List<Member> = listOf(
         memberName = GET_BUNDLE,
         descriptor = "(Lsandbox/java/lang/String;Lsandbox/java/util/Locale;Ljava/lang/ClassLoader;)Lsandbox/java/util/ResourceBundle;"
     ) {
-        override fun writeBody(emitter: EmitterModule) = with(emitter) {
+        override fun writeBody(emitter: EmitterModuleImpl) = with(emitter) {
             pushObject(0)
             pushObject(1)
             pushDefaultControl()
@@ -100,7 +100,7 @@ fun generateJavaResourceBundleMethods(): List<Member> = listOf(
         memberName = GET_BUNDLE,
         descriptor = "(Lsandbox/java/lang/String;Lsandbox/java/util/Locale;Ljava/lang/ClassLoader;Lsandbox/java/util/ResourceBundle\$Control;)Lsandbox/java/util/ResourceBundle;"
     ) {
-        override fun writeBody(emitter: EmitterModule) = with(emitter) {
+        override fun writeBody(emitter: EmitterModuleImpl) = with(emitter) {
             pushObject(0)
             pushObject(1)
             pushObject(3)
@@ -122,7 +122,7 @@ fun generateJavaResourceBundleMethods(): List<Member> = listOf(
          *     }
          *     return
          */
-        override fun writeBody(emitter: EmitterModule) = with(emitter) {
+        override fun writeBody(emitter: EmitterModuleImpl) = with(emitter) {
             lineNumber(1)
             pushObject(0)
             pushField(className, "parent", "Lsandbox/java/util/ResourceBundle;")
@@ -151,7 +151,7 @@ fun generateJavaResourceBundleMethods(): List<Member> = listOf(
          *     this.locale = locale
          *     return
          */
-        override fun writeBody(emitter: EmitterModule) = with(emitter) {
+        override fun writeBody(emitter: EmitterModuleImpl) = with(emitter) {
             pushObject(0)
             pushObject(1)
             popField(className, "name", "Lsandbox/java/lang/String;")
@@ -166,7 +166,7 @@ fun generateJavaResourceBundleMethods(): List<Member> = listOf(
     .build()
 )
 
-private fun EmitterModule.returnResourceBundle() {
+private fun EmitterModuleImpl.returnResourceBundle() {
     invokeStatic(
         owner = DJVM_NAME,
         name = GET_BUNDLE,
@@ -175,7 +175,7 @@ private fun EmitterModule.returnResourceBundle() {
     returnObject()
 }
 
-private fun EmitterModule.pushDefaultLocale() {
+private fun EmitterModuleImpl.pushDefaultLocale() {
     invokeStatic(
         owner = sandboxed(Locale::class.java),
         name = "getDefault",
@@ -183,7 +183,7 @@ private fun EmitterModule.pushDefaultLocale() {
     )
 }
 
-private fun EmitterModule.pushDefaultControl() {
+private fun EmitterModuleImpl.pushDefaultControl() {
     /*
      * The baseName parameter is expected already to have been
      * pushed onto the stack, just below the Locale value, so
