+++ djvm/src/main/kotlin/net/corda/djvm/references/Member.kt
@@ -3,12 +3,13 @@ package net.corda.djvm.references
 import net.corda.djvm.CordaInternal
 import net.corda.djvm.code.Emitter
 import net.corda.djvm.code.EmitterModule
+import java.util.function.Consumer
 
 /**
  * Alias for a handler which will replace an entire
  * method body with a block of byte-code.
  */
-typealias MethodBody = (EmitterModule) -> Unit
+typealias MethodBody = Consumer<EmitterModule>
 
 /**
  * Representation of a class member.
@@ -25,7 +26,7 @@ typealias MethodBody = (EmitterModule) -> Unit
  * @property runtimeContext Local runtime "state" objects for each [Emitter].
  */
 @CordaInternal
-data class Member(
+class Member(
     override val access: Int,
     override val className: String,
     override val memberName: String,
