+++ djvm/src/main/kotlin/net/corda/djvm/rules/implementation/ConstantFieldRemover.kt
@@ -1,9 +1,10 @@
 package net.corda.djvm.rules.implementation
 
 import net.corda.djvm.analysis.AnalysisRuntimeContext
-import net.corda.djvm.code.EmitterModule
 import net.corda.djvm.code.MemberDefinitionProvider
 import net.corda.djvm.code.impl.DJVM_NAME
+import net.corda.djvm.code.impl.EmitterModuleImpl
+import net.corda.djvm.code.impl.toMethodBody
 import net.corda.djvm.references.ImmutableMember
 import org.objectweb.asm.Opcodes.ACC_STATIC
 
@@ -22,7 +23,7 @@ object ConstantFieldRemover : MemberDefinitionProvider {
                 // This may not be needed, because this static field never
                 // had a value in the first place! Other classes load this
                 // constant value directly from the constant pool instead.
-                member.toMutable().copy(body = listOf(StringFieldInitializer(member)::writeInitializer), value = null)
+                member.toMutable().copy(body = listOf(toMethodBody(StringFieldInitializer(member)::writeInitializer)), value = null)
             } else {
                 member.toMutable().copy(value = null)
             }
@@ -35,7 +36,7 @@ object ConstantFieldRemover : MemberDefinitionProvider {
     }
 
     class StringFieldInitializer(private val member: ImmutableMember) {
-        fun writeInitializer(emitter: EmitterModule) = with(emitter) {
+        fun writeInitializer(emitter: EmitterModuleImpl) = with(emitter) {
             val value = member.value ?: return
             loadConstant(value)
             invokeStatic(DJVM_NAME, "intern", "(Ljava/lang/String;)Lsandbox/java/lang/String;", false)
