+++ djvm/src/main/kotlin/net/corda/djvm/rules/implementation/DisallowCatchingBlacklistedExceptions.kt
@@ -4,6 +4,7 @@ import net.corda.djvm.code.*
 import net.corda.djvm.code.impl.DJVM_NAME
 import net.corda.djvm.code.impl.EMIT_TRAPPING_EXCEPTIONS
 import net.corda.djvm.code.impl.THROWABLE_NAME
+import net.corda.djvm.code.impl.emit
 import net.corda.djvm.code.impl.getMemberContext
 import net.corda.djvm.code.impl.ruleViolationError
 import net.corda.djvm.code.impl.thresholdViolationError
