+++ djvm/src/main/kotlin/net/corda/djvm/rules/implementation/MemberRuleEnforcer.kt
@@ -1,15 +1,15 @@
 package net.corda.djvm.rules.implementation
 
-import net.corda.djvm.code.EmitterModule
+import net.corda.djvm.code.impl.EmitterModuleImpl
 import net.corda.djvm.references.MemberInformation
 
 class MemberRuleEnforcer(private val member: MemberInformation) {
-    fun forbidNativeMethod(emitter: EmitterModule) = with(emitter) {
+    fun forbidNativeMethod(emitter: EmitterModuleImpl) = with(emitter) {
         lineNumber(0)
         throwRuleViolationError("Native method has been deleted; ${formatFor(member)}")
     }
 
-    fun forbidAPI(emitter: EmitterModule) = with(emitter) {
+    fun forbidAPI(emitter: EmitterModuleImpl) = with(emitter) {
         lineNumber(0)
         throwRuleViolationError("Disallowed reference to API; ${formatFor(member)}")
     }
