+++ djvm/src/main/kotlin/net/corda/djvm/rules/implementation/StringConstantWrapper.kt
@@ -4,6 +4,7 @@ import net.corda.djvm.code.Emitter
 import net.corda.djvm.code.EmitterContext
 import net.corda.djvm.code.Instruction
 import net.corda.djvm.code.impl.DJVM_NAME
+import net.corda.djvm.code.impl.emit
 import net.corda.djvm.code.instructions.ConstantInstruction
 
 /**
