+++ djvm/src/main/kotlin/net/corda/djvm/rules/implementation/StubOutNativeMethods.kt
@@ -1,8 +1,9 @@
 package net.corda.djvm.rules.implementation
 
 import net.corda.djvm.analysis.AnalysisRuntimeContext
-import net.corda.djvm.code.EmitterModule
 import net.corda.djvm.code.MemberDefinitionProvider
+import net.corda.djvm.code.impl.EmitterModuleImpl
+import net.corda.djvm.code.impl.toMethodBody
 import net.corda.djvm.references.ImmutableMember
 import org.objectweb.asm.Opcodes.*
 import java.lang.reflect.Modifier
@@ -15,12 +16,18 @@ object StubOutNativeMethods : MemberDefinitionProvider {
     override fun define(context: AnalysisRuntimeContext, member: ImmutableMember): ImmutableMember = when {
         member.isMethod && isNative(member) -> member.toMutable().copy(
             access = member.access and ACC_NATIVE.inv(),
-            body = member.body + if (isForStubbing(member)) ::writeStubMethodBody else MemberRuleEnforcer(member)::forbidNativeMethod
+            body = member.body + toMethodBody(
+                if (isForStubbing(member)) {
+                    ::writeStubMethodBody
+                } else {
+                    MemberRuleEnforcer(member)::forbidNativeMethod
+                }
+            )
         )
         else -> member
     }
 
-    private fun writeStubMethodBody(emitter: EmitterModule): Unit = with(emitter) {
+    private fun writeStubMethodBody(emitter: EmitterModuleImpl): Unit = with(emitter) {
         returnVoid()
     }
