+++ djvm/src/main/kotlin/net/corda/djvm/rules/implementation/ThrowExceptionWrapper.kt
@@ -5,6 +5,7 @@ import net.corda.djvm.code.EmitterContext
 import net.corda.djvm.code.Instruction
 import net.corda.djvm.code.impl.DJVM_NAME
 import net.corda.djvm.code.impl.FROM_DJVM
+import net.corda.djvm.code.impl.emit
 import org.objectweb.asm.Opcodes.ATHROW
 
 /**
