+++ djvm/src/test/kotlin/net/corda/djvm/code/EmitterModuleTest.kt
@@ -4,6 +4,7 @@ import net.corda.djvm.SandboxType.KOTLIN
 import net.corda.djvm.TestBase
 import net.corda.djvm.analysis.impl.ClassAndMemberVisitor
 import net.corda.djvm.code.impl.ClassMutator
+import net.corda.djvm.code.impl.emit
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.objectweb.asm.ClassVisitor
